Fixed no-unused-vars eslint rule for typescript type annotation and sharp install for node 13
For the no-unused-vars we need to use the rule from typescript-eslint and not the from eslint as
that one does not now about type annotation.

For node 13 we needed to update sharp so that at least you can run yarn install and not run into
an error that sharp can not be build. This is fixed with the newer version as that has a prebuild
binary for node 13.