Added logger, fixed dist output, added new types/interfaces, removed unused imports
The new logger is mostly taken from nest framework as this does give use the change that they can implement there own logger. The logger base class has still some minor missing feature that we may still want to add like a global logger level.

The dist output was a problem with the path alias in the typescript config. The best fix was in my option to remove it as the path config is for packages like webpack and not for use when you only use tsc.

The new types are mostly for the API extensions and are added to the example so you see how to use these.