moved all modules to its own package
This is not tested at all as I have currently now working VSF instance or ES instance to test it.
It does start and he loads all modules and that all logic still works like it should I can#t
guarantee that it works like it should that needs to be tested.